consolidate all 6.8 firmware list into single list
We want to keep this around for a while to allow loading firmwares for
the currently booted kernel on major upgrade before the first reboot.
While that is generally a rather hypothetical problem, it is not
completely unrealistic and the cost of doing this are relatively low,
i.e. a few MiB of extra space.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>